% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_setting.R
\name{theme_min}
\alias{theme_min}
\title{Create mini theme for ggplot2 object.}
\usage{
theme_min(base_size = 14, face = "plain", plotTitle_just = 0.5,
  xAngle = 0, yAngle = 0, legend = "right", border_color = "grey",
  strip_fill = "gray90", grid_color = "grey80", xGrid_major = TRUE,
  yGrid_major = TRUE, xText = TRUE, yText = TRUE, tick = FALSE,
  plotBorder = NA, legend_title = TRUE)
}
\arguments{
\item{base_size}{A number for base font size.}

\item{face}{A string for fontface.}

\item{plotTitle_just}{Justment for plot title, left: 0, mid: 0.5, right: 1.}

\item{xAngle}{A number for angle of roation of X-axis text}

\item{yAngle}{A number for angle of rotation of Y-axis text}

\item{legend}{A string, oosition of lengend, one of top, bottom, left, right, none.}

\item{border_color}{A string for color of panel border.}

\item{strip_fill}{A string for color of filling strip.}

\item{grid_color}{A string for color of grid color.}

\item{xGrid_major}{Boolean, with major X-grid line or not.}

\item{yGrid_major}{Boolean, with major Y-grid line or not.}

\item{xText}{Boolean, showing X-axis text or not.}

\item{yText}{Boolean, showing Y-axis text or not.}

\item{tick}{Boolean, showing X and Y axis ticks or not.}

\item{plotBorder}{String, color of plot border.}

\item{legend_title}{Boolean, showing title of lengend or not.}
}
\description{
Create mini theme for ggplot2 object.
}
