% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot_setting.R
\name{boxplot}
\alias{boxplot}
\title{Customizing box plot}
\usage{
boxplot(fill_var = NA, fill_col = "steelblue",
  color_col = "steelblue4", alpha = 0.7, outlier_size = 1.5,
  size = 0.5)
}
\arguments{
\item{fill_var}{String, Variable name for aes coloring.}

\item{fill_col}{String, color of filling box.}

\item{color_col}{String, color of box outline and outliers}

\item{alpha}{Number, alpha of box.}

\item{outlier_size}{Number, size of outliers.}

\item{size}{Number, size of box outline.}
}
\description{
Customizing box plot
}
